Remove public `$projectDir` wherever possible
This includes all commands, though a few, where we needed to test path
existence, required the ability to inject a path, which is now handled
by a setter; when the setter is not called, `getcwd()` is then assumed.

For internal classes, this is now handled as either a passed argument,
or via a constructor argument.