Adds class for creating factory classes
Composes a `FactoryClassGenerator` in order to produce a factory for the
given class name, and then writes that class in a sibling file.

If the class is not autoloadable, `Create` raises an exception.

If unable to write the file to the filesystem, `Create` also raises an
exception.

The factory method, `createForClass`, returns the filename where the
factory class was written.