Adds class for injecting factory configuration
`ConfigInjector` creates and/or updates a special autoloadable config
file, zend-expressive-tooling-factories.global.php. Its
`injectFactoryForClass()` method expects the factory class name and the
class name to which it maps. It slurps up the existing configuration, if
any, and then adds the new mapping, sorts the entries, and writes the
file back to its location.