Ensure factory generation works in actual usage
When testing the `middleware:create` and `handler:create` commands, I
received exceptions during factory generation saying that the middleware
and/or handler class could not be found. Adding a `require $path`
statement addresses the issue, but introduces new problems during
testing, which I addressed by adding a private flag that, when disabled,
skips the `require` statement.