Provides different help messages when requesting an action
This patch modifies the `CreateHandlerCommand` to configure the command
differently based on whether an action is requested; in such a case, it
uses a completely different set of help messages and descriptions that
reference "actions" and "action classes" instead of "request handlers".

It also adds a template for the initial status message.