Ready command to accept template options
This patch updates the `CreateHandlerCommand` to do the following:

- During instantiation, it pulls the application container, and
  determines if a template renderer is registered.

- During `configure()`, if a template renderer is registered, it adds
  the following options:

  - `--without-template`, which will allow disabling generation of a
    template for the created handler.
  - `--with-template-namespace`, which allows overriding the template
    namespace used when generating the template.
  - `--with-template-name`, which allows overriding the template
    name used when generating the template.

- During `execute()`, it varies which class skeleton template to use
  based on whether or not a template renderer is registered and/or the
  `--without-template` flag is enabled. Once generation of the handler
  is done, if a template should be rendered, it is.

I had to do a bit of refactoring to make this work.

The `CreateTemplate` class renames the `forHandlerWithTemplateNamespace`
method to `generateTemplate`, and accepts an additional argument,
`$templateName`. `forHandler()` was modified to detect the template name
and pass that on to `generateTemplate()`.

The `CreateHandler` class was modified to accept the project root and
skeleton template to the constructor. The `generate()` method now
accepts the handler class name, and an optional array of additional
template substitutions, allowing the ability to generate arbitrary class
skeletons.

The `CreateHandlerCommand` class was updated to allow accepting both the
project root and the container to the constructor. This allows easier
testing capabilities. It was also updated to instantiate the
`CreateHandler` class with the new arguments, and to invoke `generate()`
using its new signature.